fix: removed utils/request.py
